+++ server.js
@@ -5,6 +5,7 @@ const userRoutes = require('./src/routes/userRoutes');
 const jobRoutes = require('./src/routes/jobRoutes');
 const applicationRoutes = require('./src/routes/applicationRoutes');
 const searchRoutes = require('./src/routes/searchRoutes');
+const bookmarkRoutes = require('./src/routes/bookmarkRoutes');
 const swaggerSpec = require('./src/config/swagger');
 const swaggerUi = require('swagger-ui-express');
 require('dotenv').config(); // dotenv 설정
@@ -39,5 +40,6 @@ app.use('/auth', userRoutes);
 app.use('/jobs', jobRoutes);
 app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerSpec)); // Swagger UI 경로
 app.use('/search', searchRoutes);
+app.use('/bookmarks', bookmarkRoutes);
 
 console.log('Swagger 문서: http://113.198.66.75:18027/api-docs');
